--- app/src/main/java/com/example/zaccianflone/pantrypro/database/Remove.java
@@ -1,31 +1,0 @@
-package com.example.zaccianflone.pantrypro.database;
-
-import com.example.zaccianflone.pantrypro.Constants;
-import com.firebase.client.DataSnapshot;
-import com.firebase.client.Firebase;
-import com.firebase.client.FirebaseError;
-import com.firebase.client.ValueEventListener;
-
-import java.util.HashMap;
-
-/**
- * Created by Ben on 7/11/2016.
- */
-public class Remove extends DatabaseUpdate {
-    //private Firebase ref = new Firebase(Constants.FIREBASE_URL);
-
-    @Override
-    public void child(String type) {
-
-    }
-
-    @Override
-    public void execute(HashMap<String, Object> values, String type) {
-
-    }
-
-    @Override
-    public void finish() {
-
-    }
-}
