--- app/src/main/java/com/example/zaccianflone/pantrypro/pantry/GroceryList.java
@@ -1,19 +1,0 @@
-package com.example.zaccianflone.pantrypro.pantry;
-
-import java.util.List;
-
-/**
- * Created by Ben on 7/12/2016.
- */
-public class GroceryList implements Pantry {
-
-    @Override
-    public void addOrEdit(List<String> items) {
-
-    }
-
-    @Override
-    public void remove(List<String> items) {
-
-    }
-}
